getError allocating only 1 byte of memory
The getError function should have been allocating EN_MAXMSG (256) bytes of memory each time it tried to access an error code.
Previously this was only 1 byte allocated so each time the function was run there was an overflow.
This was particularly causing issues when a model had negative pressure and each time step was bringing up a warning